refactor: do the _prepare_data as soon as possible
